adding dogpile,dramatiq,elasticsearch,falcon
