+++ ddtrace/contrib/dogpile_cache/__init__.py
@@ -43,8 +43,12 @@ required_modules = ["dogpile.cache"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.dogpile_cache.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.dogpile_cache.patch import get_version
+        from ..internal.dogpile_cache.patch import patch
+        from ..internal.dogpile_cache.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
