+++ ddtrace/contrib/dogpile_cache/region.py
@@ -1,55 +1,15 @@
-import dogpile
+from ddtrace.internal.utils.deprecations import DDTraceDeprecationWarning
+from ddtrace.vendor.debtcollector import deprecate
 
-from ddtrace.ext import SpanTypes
-from ddtrace.internal.constants import COMPONENT
-
-from ...constants import SPAN_MEASURED_KEY
-from ...ext import db
-from ...internal.schema import schematize_cache_operation
-from ...internal.schema import schematize_service_name
-from ...internal.utils import get_argument_value
-from ...pin import Pin
+from ..internal.dogpile_cache.region import *  # noqa: F401,F403
 
 
-def _wrap_get_create(func, instance, args, kwargs):
-    pin = Pin.get_from(dogpile.cache)
-    if not pin or not pin.enabled():
-        return func(*args, **kwargs)
+def __getattr__(name):
+    deprecate(
+        ("%s.%s is deprecated" % (__name__, name)),
+        category=DDTraceDeprecationWarning,
+    )
 
-    key = get_argument_value(args, kwargs, 0, "key")
-    with pin.tracer.trace(
-        schematize_cache_operation("dogpile.cache", cache_provider="dogpile"),
-        service=schematize_service_name(None),
-        resource="get_or_create",
-        span_type=SpanTypes.CACHE,
-    ) as span:
-        span.set_tag_str(COMPONENT, "dogpile_cache")
-        span.set_tag(SPAN_MEASURED_KEY)
-        span.set_tag("key", key)
-        span.set_tag("region", instance.name)
-        span.set_tag("backend", instance.actual_backend.__class__.__name__)
-        response = func(*args, **kwargs)
-        span.set_metric(db.ROWCOUNT, 1)
-        return response
-
-
-def _wrap_get_create_multi(func, instance, args, kwargs):
-    pin = Pin.get_from(dogpile.cache)
-    if not pin or not pin.enabled():
-        return func(*args, **kwargs)
-
-    keys = get_argument_value(args, kwargs, 0, "keys")
-    with pin.tracer.trace(
-        schematize_cache_operation("dogpile.cache", cache_provider="dogpile"),
-        service=schematize_service_name(None),
-        resource="get_or_create_multi",
-        span_type="cache",
-    ) as span:
-        span.set_tag_str(COMPONENT, "dogpile_cache")
-        span.set_tag(SPAN_MEASURED_KEY)
-        span.set_tag("keys", keys)
-        span.set_tag("region", instance.name)
-        span.set_tag("backend", instance.actual_backend.__class__.__name__)
-        response = func(*args, **kwargs)
-        span.set_metric(db.ROWCOUNT, len(response))
-        return response
+    if name in globals():
+        return globals()[name]
+    raise AttributeError("%s has no attribute %s", __name__, name)
