+++ ddtrace/contrib/dramatiq/__init__.py
@@ -35,8 +35,12 @@ required_modules = ["dramatiq"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.django.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.dramatiq.patch import get_version
+        from ..internal.dramatiq.patch import patch
+        from ..internal.dramatiq.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
