+++ ddtrace/contrib/elasticsearch/__init__.py
@@ -49,9 +49,9 @@ Example::
     # Override service name
     config.elasticsearch['service'] = 'custom-service-name'
 """
-from .patch import get_version
-from .patch import get_versions
-from .patch import patch
+from ..internal.elasticsearch.patch import get_version
+from ..internal.elasticsearch.patch import get_versions
+from ..internal.elasticsearch.patch import patch
 
 
 __all__ = ["patch", "get_version", "get_versions"]
