+++ ddtrace/contrib/falcon/__init__.py
@@ -51,8 +51,10 @@ required_modules = ["falcon"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .middleware import TraceMiddleware
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.falcon.patch` directly
+        from . import patch as _  # noqa: F401, I001
+        from ..internal.falcon.middleware import TraceMiddleware
+        from ..internal.falcon.patch import get_version
+        from ..internal.falcon.patch import patch
 
         __all__ = ["TraceMiddleware", "patch", "get_version"]
